feat: Creating the Settings Page
