+++ src/components/Profile/components/Notifications/Notifications.module.scss
@@ -1,4 +1,5 @@
 .container {
+  margin-top: 3rem;
   padding: 1rem;
   border: 1px solid #ccc;
   border-radius: 8px;
