+++ src/components/Profile/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./DeleteAccount";
+export * from "./Notifications";
 export * from "./UpdatePassword";
