+++ src/components/Profile/types/index.ts
@@ -1,0 +1,4 @@
+import { PROFILE_STATUS } from "../constants";
+
+export type ProfileStatus =
+  (typeof PROFILE_STATUS)[keyof typeof PROFILE_STATUS];
