+++ src/components/index.ts
@@ -3,4 +3,3 @@ export * from "./NotFound";
 export * from "./Seo";
 export * from "./PostCard";
 export * from "./PrivateRoute";
-export * from "./Notifications";
