+++ src/constants/index.ts
@@ -1,3 +1,2 @@
 export * from "./languages";
-export * from "./status";
 export * from "./url";
